[DF-13309] formulas: review fixes
Former-commit-id: 82c9480bf6f3674492d5f9e08d475d3f06216b4f

+++ Sources/Internal/UI/Formula/Private/FormulaTokenizer.cpp
@@ -469,7 +469,7 @@ bool FormulaTokenizer::IsDigit(char ch) const
 void FormulaTokenizer::ReadChar()
 {
     currentPosition++;
-    if (currentPosition < str.size())
+    if (currentPosition < static_cast<int32>(str.size()))
     {
         ch = str.at(currentPosition);
         positionInLine++;
@@ -477,7 +477,7 @@ void FormulaTokenizer::ReadChar()
     else
     {
         ch = '\0';
-        currentPosition = str.size();
+        currentPosition = static_cast<int32>(str.size());
     }
 }
 }
